requirements.txt
allure-pytest
pytest==7.2.2
requests
arrow
polling
pygsheets
gspread
retry